+++ go2_robot_sdk/scripts_go2/go2_camerainfo.py
