+++ go2_robot_sdk/scripts_go2/webrtc_driver.py
@@ -39,7 +39,7 @@ import requests
 from aiortc import RTCPeerConnection, RTCSessionDescription
 
 
-from scripts.go2_lidar_decoder import LidarDecoder
+from scripts_go2.go2_lidar_decoder import LidarDecoder
 
 
 logging.basicConfig(level=logging.WARN)
